[INJIMOB-691] rename file backupAndRestore to backupAndRestoreSetup
